+++ corehq/ex-submodules/couchforms/tests/__init__.py
@@ -12,8 +12,8 @@ try:
     from .test_post import *
     from .test_xml import *
     from .test_errors import *
-    from .test_adjust_datetimes import *
-    from .test_dbaccessors import *
+    #from .test_adjust_datetimes import *
+    #from .test_dbaccessors import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
