+++ corehq/ex-submodules/couchforms/tests/test_errors.py
@@ -1,9 +1,11 @@
 from django.test import TestCase
 from corehq.apps.receiverwrapper import submit_form_locally
-from couchforms.models import XFormError
+
+from corehq.form_processor.interfaces import FormProcessorInterface
 
 
 class CaseProcessingErrorsTest(TestCase):
+
     def test_no_case_id(self):
         """
         submit form with a case block that has no case_id
@@ -31,19 +33,13 @@ class CaseProcessingErrorsTest(TestCase):
             </data>""",
             'my_very_special_domain',
         )
-        xform_errors = XFormError.view(
-            'domain/docs',
-            startkey=['my_very_special_domain', 'XFormError'],
-            endkey=['my_very_special_domain', 'XFormError', {}],
-            reduce=False,
-            include_docs=True,
-        ).all()
+        xform_errors = FormProcessorInterface.get_by_doc_type('my_very_special_domain', 'XFormError')
 
         related_errors = [xform_error for xform_error in xform_errors
-                          if xform_error.get_id == 'abc-easy-as-123']
+                          if xform_error.to_generic().id == 'abc-easy-as-123']
         self.assertEqual(len(related_errors), 1)
         related_error = related_errors[0]
-        self.assertEqual(related_error.problem,
+        self.assertEqual(related_error.to_generic().problem,
                          'IllegalCaseId: case_id must not be empty')
 
     def test_uses_referrals(self):
@@ -71,17 +67,11 @@ class CaseProcessingErrorsTest(TestCase):
             </data>""",
             'my_very_special_domain',
         )
-        xform_errors = XFormError.view(
-            'domain/docs',
-            startkey=['my_very_special_domain', 'XFormError'],
-            endkey=['my_very_special_domain', 'XFormError', {}],
-            reduce=False,
-            include_docs=True,
-        ).all()
+        xform_errors = FormProcessorInterface.get_by_doc_type('my_very_special_domain', 'XFormError')
 
         related_errors = [xform_error for xform_error in xform_errors
-                          if xform_error.get_id == 'abc-easy-as-456']
+                          if xform_error.to_generic().id == 'abc-easy-as-456']
         self.assertEqual(len(related_errors), 1)
         related_error = related_errors[0]
-        self.assertEqual(related_error.problem,
+        self.assertEqual(related_error.to_generic().problem,
                 'UsesReferrals: Sorry, referrals are no longer supported!')
