--- corehq/form_processor/tests/data/date_in_meta.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<data xmlns="http://bihar.commcarehq.org/mcts/create_mcts_persona">
-    <meta xmlns="http://openrosa.org/jr/xforms">
-        <instanceID>c28f68a5-e920-4594-aea7-dbae3c3650b4</instanceID>
-        <timeEnd>2014-07-11</timeEnd>
-        <timeStart>2014-07-10</timeStart>
-        <userID>12345</userID>
-    </meta>
-    <foo>bar</foo>
-</data>
