--- corehq/form_processor/tests/data/decimalmeta.json
@@ -1,79 +1,0 @@
-{
-    "_attachments": {
-        "form.xml": {
-            "content_type": "text/xml",
-            "digest": "md5-Ii5K6MOVhQTPZWrGsYUGSQ==",
-            "length": 1183,
-            "revpos": 1,
-            "stub": true
-        }
-    },
-    "_id": "7e5d13569f6b3d71c4025a624d4a4aff",
-    "_rev": "1-3d7a2b54d120fbacb5df672503bc2eb8",
-    "computed_": {},
-    "computed_modified_on_": null,
-    "doc_type": "XFormInstance",
-    "form": {
-        "#type": "general",
-        "@xmlns": "http://openrosa.org/app/general",
-        "assessment": {
-            "categories": "other"
-        },
-        "case": {
-            "case_type": "malaria",
-            "followup_date": "",
-            "followup_type": "referral",
-            "outcome": "",
-            "patient_id": "6713c1c211b3af6832c8ad781e310d98"
-        },
-        "chw_referral_id": "",
-        "danger_signs": "convuls",
-        "diagnosis": "malaria",
-        "drugs": {
-            "dispensed": "",
-            "prescribed": ""
-        },
-        "encounter_date": "2010-07-22",
-        "encounter_type": "new_case",
-        "general_exam": "none",
-        "hiv_result": "r",
-        "investigations": {
-            "categories": ""
-        },
-        "medical_history": "daibetes",
-        "meta": {
-            "appVersion": "2.0",
-            "clinic_id": "5020280",
-            "timeEnd": "2010-07-23T13:55:11.648000Z",
-            "timeStart": "2010-07-22T13:54:27.971000Z",
-            "userID": "f7f0c79e-8b79-11df-b7de-005056c00008",
-            "username": "admin"
-        },
-        "on_haart": "",
-        "phys_exam_detail": "none",
-        "resolution": "referral",
-        "vitals": {
-            "bp": "",
-            "heart_rate": "",
-            "height": "",
-            "resp_rate": "",
-            "temp": "",
-            "weight": ""
-        }
-    },
-    "history": [],
-    "initial_processing_complete": false,
-    "partial_submission": false,
-    "received_on": "2013-09-21T04:06:18Z",
-    "xmlns": "http://openrosa.org/app/general",
-    "#export_tag": null,
-    "app_id": null,
-    "auth_context": {},
-    "build_id": null,
-    "date_header": null,
-    "domain": null,
-    "last_sync_token": null,
-    "openrosa_headers": {},
-    "path": null,
-    "submit_ip": null
-}
