--- corehq/form_processor/tests/data/deprecation/edit.xml
@@ -1,17 +1,0 @@
-<?xml version='1.0' ?>
-<general xmlns="http://openrosa.org/app/general">
-	<Meta>
-		<TimeStart>2010-07-22T13:54:27.971</TimeStart>
-		<TimeEnd>2010-07-24T13:55:11.648</TimeEnd>
-		<username>admin</username>
-		<uid>7H46J37FGH3</uid>
-        <user_id>f7f0c79e-8b79-11df-b7de-005056c00008</user_id>
-	</Meta>
-	<encounter_date>2010-07-22</encounter_date>
-    <vitals>
-        <height>100</height>
-    </vitals>
-    <assessment>
-        <categories>Edited Baby!</categories>
-    </assessment>
-</general>
