--- corehq/form_processor/tests/data/gps_empty_location.xml
@@ -1,14 +1,0 @@
-<?xml version='1.0' ?>
-<data uiVersion="1" version="7" name="New Form" xmlns:jrm="http://dev.commcarehq.org/jr/xforms" xmlns="http://openrosa.org/formdesigner/901A0E79-BD35-4211-9212-7D14442ED8B2">
-<something>Allegro</something>
-<n0:meta xmlns:n0="http://openrosa.org/jr/xforms">
-<n0:deviceID>355031040087328</n0:deviceID>
-<n0:timeStart>2014-08-13T14:39:01.459-04</n0:timeStart>
-<n0:timeEnd>2014-08-13T14:39:54.914-04</n0:timeEnd>
-<n0:username>jenny</n0:username>
-<n0:userID>1defd60e1fa13787cd5dd208daf48817</n0:userID>
-<n0:instanceID>3c015454-993d-4061-b56b-c324e8f2814b</n0:instanceID>
-<n1:location xmlns:n1="http://commcarehq.org/xforms" />
-<n2:appVersion xmlns:n2="http://commcarehq.org/xforms">CommCare ODK, version "2.15"(106). App v18. CommCare Version 2.15. Build CUSTOM_BUILD, built on: UNKNOWN_DATE</n2:appVersion>
-</n0:meta>
-</data>
