--- corehq/form_processor/tests/data/meta_bad_username.json
@@ -1,69 +1,0 @@
-{
-    "_attachments": {
-        "form.xml": {
-            "content_type": "text/xml",
-            "digest": "md5-KjAfU4FAA8utAFliTkG5xA==",
-            "length": 2032,
-            "revpos": 1,
-            "stub": true
-        }
-    },
-    "_id": "e8afaec3c66745ef80e48062d4b91b56",
-    "_rev": "1-04d5688db3141f16c94cc87b3871b848",
-    "computed_": {},
-    "computed_modified_on_": null,
-    "doc_type": "XFormInstance",
-    "form": {
-        "#type": "data",
-        "@uiVersion": "1",
-        "@version": "200",
-        "@xmlns": "http://openrosa.org/formdesigner/badmetauser",
-        "contact_phone_number": "",
-        "contact_phone_number_is_verified": "",
-        "meta": {
-            "@xmlns": "http://openrosa.org/jr/xforms",
-            "appVersion": {
-                "#text": "2.0",
-                "@xmlns": "http://commcarehq.org/xforms"
-            },
-            "deviceID": "commconnect",
-            "instanceID": "e8afaec3c66745ef80e48062d4b91b56",
-            "timeEnd": "2013-07-20T00:02:27.493000Z",
-            "timeStart": "2013-07-19T21:21:31.188000Z",
-            "userID": "f7f0c79e-8b79-11df-b7de-005056c00008",
-            "username": "2013-07-19"
-        },
-        "q10_participant_read_and_write_portuguese": "",
-        "q11_can_participant_receive_and_send_sms_messages": "\n    ",
-        "q12_participant_read_informed_consent_and_agree_to_participant": "\n    ",
-        "q2_participant_gender": "",
-        "q3_participant_age_in_years": "",
-        "q4_participant_pregnant_or_has_child": "",
-        "q5_participant_currently_using_cm": "",
-        "q6_participant_current_contraceptive_method": "",
-        "q7_plan_to_delay_pregnancy_next_year": "",
-        "q8_participant_and_partner_use_condom_during_intercourse": "\n    ",
-        "q9_participant_have_sim_with_dedicated_number_and_mobile_phone_access": "\n    ",
-        "reg_date": "2013-09-01",
-        "reg_date_day_of_week": "",
-        "reg_date_now": "2013-09-01",
-        "registration_confirmation": "",
-        "reminder_start_date": "2013-09-24",
-        "start_offset": ""
-    },
-    "history": [],
-    "initial_processing_complete": false,
-    "partial_submission": false,
-    "received_on": "2013-09-21T04:16:59Z",
-    "xmlns": "http://openrosa.org/formdesigner/badmetauser",
-    "#export_tag": null,
-    "app_id": null,
-    "auth_context": {},
-    "build_id": null,
-    "date_header": null,
-    "domain": null,
-    "last_sync_token": null,
-    "openrosa_headers": {},
-    "path": null,
-    "submit_ip": null
-}
