--- corehq/form_processor/tests/data/unicode.json
@@ -1,36 +1,0 @@
-{
-    "_attachments": {
-        "form.xml": {
-            "content_type": "text/xml",
-            "digest": "md5-6gRAwanOCwhm48+WGlV8lA==",
-            "length": 127,
-            "revpos": 1,
-            "stub": true
-        }
-    },
-    "_id": "0a15b26d3c8d62eb9436a045bdbdc379",
-    "_rev": "1-ce760643ddecd48e71e912bf8f2ad9d7",
-    "computed_": {},
-    "computed_modified_on_": null,
-    "doc_type": "XFormInstance",
-    "form": {
-        "#type": "unicode",
-        "@xmlns": "http://openrosa.org/app/unicode",
-        "datum": "\u7edf\u4e00\u7801"
-    },
-    "history": [],
-    "initial_processing_complete": false,
-    "partial_submission": false,
-    "received_on": "2014-01-06T21:30:43Z",
-    "xmlns": "http://openrosa.org/app/unicode",
-    "#export_tag": null,
-    "app_id": null,
-    "auth_context": {},
-    "build_id": null,
-    "date_header": null,
-    "domain": null,
-    "last_sync_token": null,
-    "openrosa_headers": {},
-    "path": null,
-    "submit_ip": null
-}
