--- corehq/form_processor/tests/data/unicode.xml
@@ -1,4 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<unicode xmlns="http://openrosa.org/app/unicode">
-	<datum>统一码</datum>
-</unicode>
