--- corehq/form_processor/tests/test_duplicates.py
@@ -1,56 +1,0 @@
-import os
-from django.test import TestCase
-from couchforms.models import XFormInstance
-
-from ..interfaces import FormProcessorInterface
-from ..generic import GenericXFormInstance
-
-
-class DuplicateFormTest(TestCase):
-    ID = '7H46J37FGH3'
-
-    def tearDown(self):
-        XFormInstance.get_db().flush()
-
-    def _get_file(self):
-        file_path = os.path.join(os.path.dirname(__file__), "data", "duplicate.xml")
-        with open(file_path, "rb") as f:
-            return f.read()
-
-    def test_basic_duplicate(self):
-        xml_data = self._get_file()
-        xform = FormProcessorInterface.post_xform(xml_data)
-        self.assertEqual(self.ID, xform.to_generic().id)
-        self.assertEqual("XFormInstance", xform.to_generic().doc_type)
-        self.assertEqual("test-domain", xform.to_generic().domain)
-
-        xform = FormProcessorInterface.post_xform(xml_data, domain='test-domain')
-        self.assertNotEqual(self.ID, xform.to_generic().id)
-        self.assertEqual("XFormDuplicate", xform.to_generic().doc_type)
-        self.assertTrue(self.ID in xform.to_generic().problem)
-
-    def test_wrong_doc_type(self):
-        domain = 'test-domain'
-        generic_xform = GenericXFormInstance(
-            doc_type='Foo',
-            domain=domain,
-        )
-        xform = FormProcessorInterface.create_from_generic(generic_xform)
-
-        instance = self._get_file()
-        instance = instance.replace(self.ID, xform.to_generic().id)
-        xform = FormProcessorInterface.post_xform(instance, domain=domain)
-        self.assertNotEqual(xform.to_generic().id, self.ID)
-
-    def test_wrong_domain(self):
-        domain = 'test-domain'
-        generic_xform = GenericXFormInstance(
-            doc_type='XFormInstance',
-            domain='wrong-domain',
-        )
-        xform = FormProcessorInterface.create_from_generic(generic_xform)
-
-        instance = self._get_file()
-        instance = instance.replace(self.ID, xform.to_generic().id)
-        xform = FormProcessorInterface.post_xform(instance, domain=domain)
-        self.assertNotEqual(xform.to_generic().id, self.ID)
